requirements.txt
flask
google-genai
python-dotenv
langchain_community
langchain
